api/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.104.1
h11==0.14.0
httptools==0.6.1
idna==3.6
mysql-connector-python==8.2.0
protobuf==4.21.12
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
python-multipart==0.0.6
python-slugify==8.0.1
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
text-unidecode==1.3
typing_extensions==4.8.0
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
